+++ src/commands/runpr.rs
@@ -1,4 +1,5 @@
 use crate::github::{CreateLabel, CreatePullRequest, GitHub};
+use crate::logging::{logfailures, logproject};
 use crate::project::Project;
 use crate::util::{Options, RunOpts, Runner};
 use clap::Args;
@@ -89,7 +90,7 @@ impl RunPr {
             if github.get_repository(ghrepo)?.archived {
                 continue;
             }
-            boldln!("{}", p.name());
+            logproject(&p);
             let defbranch = p.default_branch()?;
             p.stash(opts.quiet)?;
             p.runcmd("git")
@@ -107,7 +108,7 @@ impl RunPr {
             // XXX: When adding support for commands that commit, also check
             //      whether $branch is ahead of $defbranch.
             if !p.has_staged_changes()? {
-                println!("> No changes"); // TODO: Style output?
+                log::info!("No changes");
                 p.runcmd("git")
                     .arg("checkout")
                     .arg(defbranch)
@@ -158,7 +159,7 @@ impl RunPr {
                                 description: None,
                             },
                         )?;
-                        println!("> Created label {lbl:?} in {ghrepo}"); // TODO: Style output?
+                        log::info!("Created label {lbl:?} in {ghrepo}");
                     }
                     labels.push(lbl.as_str());
                 }
@@ -172,12 +173,7 @@ impl RunPr {
                 }
             }
         }
-        if !failures.is_empty() {
-            boldln!("\nFailures:");
-            for p in failures {
-                println!("{}", p.name());
-            }
-        }
+        logfailures(failures);
         Ok(())
     }
 }