+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v0.3.0 (in development)
 - Added a short `-R` form for `--root`
 - Added `--stash` option to `run` command
 - Added `--label` and `--soft-label` options to the `run-pr` command
+- Added `--verbose` option
 
 v0.2.0 (2025-01-06)
 -------------------